datastructures/get_node_value.go
// https://www.hackerrank.com/challenges/get-the-value-of-the-node-at-a-specific-position-from-the-tail/problem?isFullScreen=true

package datastructures

type SinglyLinkedListNode struct {
	data int32
	next *SinglyLinkedListNode
}

type SinglyLinkedList struct {
	head *SinglyLinkedListNode
	tail *SinglyLinkedListNode
}

func (singlyLinkedList *SinglyLinkedList) insertNodeIntoSinglyLinkedList(nodeData int32) {
	node := &SinglyLinkedListNode{
		next: nil,
		data: nodeData,
	}

	if singlyLinkedList.head == nil {
		singlyLinkedList.head = node
	} else {
		singlyLinkedList.tail.next = node
	}

	singlyLinkedList.tail = node
}

func getNode(node *SinglyLinkedListNode, positionFromTail int32) int32 {
	length := 0
	n := node.next

	for n.next != nil {
		length++
		n = n.next
	}

	n = node.next
	for i := 0; i < length-int(positionFromTail)+1; i++ {
		n = n.next
	}

	return n.data
}
